Create angular_index.html
Here is file 1 of 2 of a neat little feature. This is the html for a very basic todo list app.  File 2 of 2 is the app.js file.  Credit for this code goes to Era Balliu of Web Code Geeks, who wrote this article.  Here is the link: https://www.webcodegeeks.com/javascript/angular-js/angular-js-todo-app-example/