Create angular_app.js
Here is file 2 of 2.  A very basic todo list app, written in html and angularJS.  The purpose of posting this is to provide inspiration to get everybody to start writing code for our club forum!  Credit for this code goes to Era Balliu of Web Code Geeks. https://www.webcodegeeks.com/javascript/angular-js/angular-js-todo-app-example/